Created file sympy/assumptions/CNF.py to hold the code related to the generation and handling of CNF objects
