+++ sympy/assumptions/ask_generated.py
@@ -7,85 +7,86 @@ Instead, run ./bin/ask_update.py.
 """
 
 from sympy.core.cache import cacheit
-from sympy.logic.boolalg import And, Literal
+from sympy.logic.boolalg import And
+from sympy.assumptions.CNF import Literal
 from sympy.assumptions.ask import Q
 
 # -{ Known facts as a set }-
 @cacheit
 def get_all_known_facts():
     return {
-        frozenset({Literal( Q.diagonal, True), Literal( Q.upper_triangular, False)}),
-        frozenset({Literal( Q.infinite, True), Literal( Q.extended_real, False)}),
-        frozenset({Literal( Q.square, False), Literal( Q.invertible, True)}),
-        frozenset({Literal( Q.lower_triangular, True), Literal( Q.triangular, False)}),
-        frozenset({Literal( Q.nonpositive, False), Literal( Q.negative, True)}),
-        frozenset({Literal( Q.integer_elements, True), Literal( Q.real_elements, False)}),
-        frozenset({Literal( Q.orthogonal, True), Literal( Q.unitary, False)}),
-        frozenset({Literal( Q.positive, True), Literal( Q.negative, True)}),
-        frozenset({Literal( Q.algebraic, False), Literal( Q.transcendental, False), Literal( Q.finite, True), Literal( Q.complex, True)}),
-        frozenset({Literal( Q.real, True), Literal( Q.complex, False)}),
-        frozenset({Literal( Q.square, False), Literal( Q.symmetric, True)}),
-        frozenset({Literal( Q.infinite, True), Literal( Q.finite, True)}),
-        frozenset({Literal( Q.zero, True), Literal( Q.even, False)}),
-        frozenset({Literal( Q.nonzero, False), Literal( Q.negative, True)}),
-        frozenset({Literal( Q.orthogonal, False), Literal( Q.real, True), Literal( Q.unitary, True)}),
-        frozenset({Literal( Q.complex, False), Literal( Q.transcendental, True)}),
-        frozenset({Literal( Q.real, True), Literal( Q.hermitian, False)}),
-        frozenset({Literal( Q.odd, True), Literal( Q.integer, False)}),
-        frozenset({Literal( Q.square, False), Literal( Q.normal, True)}),
-        frozenset({Literal( Q.positive_definite, True), Literal( Q.invertible, False)}),
-        frozenset({Literal( Q.nonnegative, True), Literal( Q.positive, False), Literal( Q.zero, False)}),
-        frozenset({Literal( Q.finite, False), Literal( Q.algebraic, True)}),
-        frozenset({Literal( Q.positive, True), Literal( Q.zero, True)}),
-        frozenset({Literal( Q.zero, True), Literal( Q.nonpositive, False)}),
-        frozenset({Literal( Q.negative, False), Literal( Q.nonpositive, True), Literal( Q.zero, False)}),
-        frozenset({Literal( Q.diagonal, True), Literal( Q.normal, False)}),
-        frozenset({Literal( Q.rational, True), Literal( Q.irrational, True)}),
-        frozenset({Literal( Q.triangular, False), Literal( Q.unit_triangular, True)}),
-        frozenset({Literal( Q.positive, False), Literal( Q.prime, True)}),
-        frozenset({Literal( Q.unitary, True), Literal( Q.invertible, False)}),
-        frozenset({Literal( Q.finite, False), Literal( Q.transcendental, True)}),
-        frozenset({Literal( Q.real, False), Literal( Q.zero, True)}),
-        frozenset({Literal( Q.normal, False), Literal( Q.unitary, True)}),
-        frozenset({Literal( Q.nonnegative, False), Literal( Q.zero, True)}),
-        frozenset({Literal( Q.positive, True), Literal( Q.nonnegative, False)}),
-        frozenset({Literal( Q.hermitian, True), Literal( Q.antihermitian, True)}),
-        frozenset({Literal( Q.zero, True), Literal( Q.negative, True)}),
-        frozenset({Literal( Q.positive, False), Literal( Q.negative, False), Literal( Q.nonzero, True)}),
         frozenset({Literal( Q.positive_definite, False), Literal( Q.orthogonal, True)}),
-        frozenset({Literal( Q.transcendental, True), Literal( Q.algebraic, True)}),
-        frozenset({Literal( Q.positive, True), Literal( Q.nonzero, False)}),
-        frozenset({Literal( Q.lower_triangular, False), Literal( Q.diagonal, True)}),
-        frozenset({Literal( Q.fullrank, False), Literal( Q.invertible, True)}),
-        frozenset({Literal( Q.fullrank, True), Literal( Q.invertible, False), Literal( Q.square, True)}),
-        frozenset({Literal( Q.real, False), Literal( Q.irrational, True)}),
-        frozenset({Literal( Q.integer, False), Literal( Q.prime, True)}),
-        frozenset({Literal( Q.rational, True), Literal( Q.real, False)}),
-        frozenset({Literal( Q.invertible, False), Literal( Q.singular, False)}),
+        frozenset({Literal( Q.orthogonal, True), Literal( Q.unitary, False)}),
+        frozenset({Literal( Q.algebraic, True), Literal( Q.transcendental, True)}),
+        frozenset({Literal( Q.positive, True), Literal( Q.zero, True)}),
+        frozenset({Literal( Q.nonnegative, False), Literal( Q.zero, True)}),
         frozenset({Literal( Q.triangular, False), Literal( Q.upper_triangular, True)}),
-        frozenset({Literal( Q.positive, True), Literal( Q.real, False)}),
-        frozenset({Literal( Q.singular, True), Literal( Q.invertible, True)}),
-        frozenset({Literal( Q.real, True), Literal( Q.rational, False), Literal( Q.irrational, False), Literal( Q.finite, True)}),
-        frozenset({Literal( Q.odd, True), Literal( Q.even, True)}),
-        frozenset({Literal( Q.rational, True), Literal( Q.finite, False)}),
-        frozenset({Literal( Q.integer, False), Literal( Q.even, True)}),
+        frozenset({Literal( Q.zero, False), Literal( Q.nonpositive, True), Literal( Q.negative, False)}),
+        frozenset({Literal( Q.finite, True), Literal( Q.infinite, True)}),
+        frozenset({Literal( Q.irrational, True), Literal( Q.rational, True)}),
+        frozenset({Literal( Q.invertible, True), Literal( Q.square, False)}),
         frozenset({Literal( Q.finite, False), Literal( Q.irrational, True)}),
-        frozenset({Literal( Q.infinite, False), Literal( Q.real, False), Literal( Q.extended_real, True)}),
-        frozenset({Literal( Q.real_elements, True), Literal( Q.complex_elements, False)}),
-        frozenset({Literal( Q.lower_triangular, False), Literal( Q.upper_triangular, False), Literal( Q.triangular, True)}),
-        frozenset({Literal( Q.complex, False), Literal( Q.algebraic, True)}),
-        frozenset({Literal( Q.rational, True), Literal( Q.algebraic, False)}),
-        frozenset({Literal( Q.real, True), Literal( Q.zero, False), Literal( Q.negative, False), Literal( Q.positive, False)}),
-        frozenset({Literal( Q.complex, False), Literal( Q.imaginary, True)}),
-        frozenset({Literal( Q.composite, True), Literal( Q.prime, True)}),
-        frozenset({Literal( Q.real, True), Literal( Q.imaginary, True)}),
-        frozenset({Literal( Q.diagonal, False), Literal( Q.lower_triangular, True), Literal( Q.upper_triangular, True)}),
-        frozenset({Literal( Q.diagonal, True), Literal( Q.symmetric, False)}),
+        frozenset({Literal( Q.zero, False), Literal( Q.nonnegative, True), Literal( Q.positive, False)}),
+        frozenset({Literal( Q.positive, True), Literal( Q.real, False)}),
+        frozenset({Literal( Q.antihermitian, False), Literal( Q.imaginary, True)}),
+        frozenset({Literal( Q.orthogonal, False), Literal( Q.real, True), Literal( Q.unitary, True)}),
+        frozenset({Literal( Q.complex_elements, False), Literal( Q.real_elements, True)}),
+        frozenset({Literal( Q.square, True), Literal( Q.fullrank, True), Literal( Q.invertible, False)}),
+        frozenset({Literal( Q.nonpositive, False), Literal( Q.negative, True)}),
+        frozenset({Literal( Q.upper_triangular, False), Literal( Q.lower_triangular, False), Literal( Q.triangular, True)}),
         frozenset({Literal( Q.rational, False), Literal( Q.integer, True)}),
-        frozenset({Literal( Q.odd, False), Literal( Q.even, False), Literal( Q.integer, True)}),
-        frozenset({Literal( Q.real, True), Literal( Q.extended_real, False)}),
+        frozenset({Literal( Q.negative, True), Literal( Q.zero, True)}),
+        frozenset({Literal( Q.algebraic, True), Literal( Q.complex, False)}),
+        frozenset({Literal( Q.invertible, True), Literal( Q.fullrank, False)}),
+        frozenset({Literal( Q.integer, False), Literal( Q.odd, True)}),
+        frozenset({Literal( Q.normal, True), Literal( Q.square, False)}),
+        frozenset({Literal( Q.invertible, False), Literal( Q.positive_definite, True)}),
+        frozenset({Literal( Q.negative, False), Literal( Q.nonzero, True), Literal( Q.positive, False)}),
+        frozenset({Literal( Q.real, False), Literal( Q.zero, True)}),
+        frozenset({Literal( Q.symmetric, True), Literal( Q.square, False)}),
+        frozenset({Literal( Q.prime, True), Literal( Q.positive, False)}),
+        frozenset({Literal( Q.real, True), Literal( Q.hermitian, False)}),
+        frozenset({Literal( Q.complex, True), Literal( Q.transcendental, False), Literal( Q.algebraic, False), Literal( Q.finite, True)}),
+        frozenset({Literal( Q.invertible, False), Literal( Q.unitary, True)}),
         frozenset({Literal( Q.real, False), Literal( Q.negative, True)}),
-        frozenset({Literal( Q.antihermitian, False), Literal( Q.imaginary, True)})
+        frozenset({Literal( Q.symmetric, False), Literal( Q.diagonal, True)}),
+        frozenset({Literal( Q.lower_triangular, True), Literal( Q.upper_triangular, True), Literal( Q.diagonal, False)}),
+        frozenset({Literal( Q.real, False), Literal( Q.irrational, True)}),
+        frozenset({Literal( Q.positive, True), Literal( Q.nonnegative, False)}),
+        frozenset({Literal( Q.real, True), Literal( Q.extended_real, False)}),
+        frozenset({Literal( Q.even, True), Literal( Q.odd, True)}),
+        frozenset({Literal( Q.finite, False), Literal( Q.transcendental, True)}),
+        frozenset({Literal( Q.nonzero, False), Literal( Q.negative, True)}),
+        frozenset({Literal( Q.zero, True), Literal( Q.even, False)}),
+        frozenset({Literal( Q.zero, False), Literal( Q.negative, False), Literal( Q.real, True), Literal( Q.positive, False)}),
+        frozenset({Literal( Q.normal, False), Literal( Q.diagonal, True)}),
+        frozenset({Literal( Q.real_elements, False), Literal( Q.integer_elements, True)}),
+        frozenset({Literal( Q.prime, True), Literal( Q.integer, False)}),
+        frozenset({Literal( Q.normal, False), Literal( Q.unitary, True)}),
+        frozenset({Literal( Q.triangular, False), Literal( Q.unit_triangular, True)}),
+        frozenset({Literal( Q.rational, False), Literal( Q.irrational, False), Literal( Q.real, True), Literal( Q.finite, True)}),
+        frozenset({Literal( Q.upper_triangular, False), Literal( Q.diagonal, True)}),
+        frozenset({Literal( Q.imaginary, True), Literal( Q.complex, False)}),
+        frozenset({Literal( Q.positive, True), Literal( Q.nonzero, False)}),
+        frozenset({Literal( Q.hermitian, True), Literal( Q.antihermitian, True)}),
+        frozenset({Literal( Q.real, True), Literal( Q.complex, False)}),
+        frozenset({Literal( Q.triangular, False), Literal( Q.lower_triangular, True)}),
+        frozenset({Literal( Q.invertible, True), Literal( Q.singular, True)}),
+        frozenset({Literal( Q.finite, False), Literal( Q.algebraic, True)}),
+        frozenset({Literal( Q.real, False), Literal( Q.rational, True)}),
+        frozenset({Literal( Q.infinite, True), Literal( Q.extended_real, False)}),
+        frozenset({Literal( Q.even, True), Literal( Q.integer, False)}),
+        frozenset({Literal( Q.integer, True), Literal( Q.even, False), Literal( Q.odd, False)}),
+        frozenset({Literal( Q.positive, True), Literal( Q.negative, True)}),
+        frozenset({Literal( Q.imaginary, True), Literal( Q.real, True)}),
+        frozenset({Literal( Q.transcendental, True), Literal( Q.complex, False)}),
+        frozenset({Literal( Q.finite, False), Literal( Q.rational, True)}),
+        frozenset({Literal( Q.invertible, False), Literal( Q.singular, False)}),
+        frozenset({Literal( Q.algebraic, False), Literal( Q.rational, True)}),
+        frozenset({Literal( Q.nonpositive, False), Literal( Q.zero, True)}),
+        frozenset({Literal( Q.infinite, False), Literal( Q.real, False), Literal( Q.extended_real, True)}),
+        frozenset({Literal( Q.lower_triangular, False), Literal( Q.diagonal, True)}),
+        frozenset({Literal( Q.prime, True), Literal( Q.composite, True)})
     }
 
 # -{ Known facts in Conjunctive Normal Form }-
