+++ sympy/logic/algorithms/dpll2.py
@@ -14,8 +14,8 @@ from collections import defaultdict
 from heapq import heappush, heappop
 
 from sympy.core.compatibility import range
-from sympy import default_sort_key, ordered
-from sympy.logic.boolalg import conjuncts, to_cnf, to_int_repr, _find_predicates, CNF, EncodedCNF
+from sympy import ordered
+from sympy.assumptions.CNF import EncodedCNF
 
 
 def dpll_satisfiable(expr, all_models=False):
