+++ sympy/logic/boolalg.py
@@ -2579,370 +2579,3 @@ def simplify_patterns_xor():
                       And(Lt(a, Max(b, c)), Ge(a, Min(b, c)))),
                      )
     return _matchers_xor
-
-
-literals_store = dict()  # A strategy to reduce extra creation of Literals
-
-
-class Literal(object):
-    """
-    The smallest element of a CNF object
-    """
-    def __new__(cls, lit, is_not=False):
-        if type(lit) is int:
-            obj = super(Literal, cls).__new__(cls)
-            obj.code = lit
-            obj.is_Not = is_not
-            return obj
-        if type(lit).__name__ == 'Not':
-            lit = lit.args[0]
-            is_not = True
-        k = literals_store.get(lit, None)
-        if k is not None:
-            return Literal(k, is_not)
-        else:
-            obj = super(Literal, cls).__new__(cls)
-            obj.code = len(literals_store) + 1
-            literals_store[lit] = obj.code
-            obj.is_Not = is_not
-            return obj
-
-    @property
-    def arg(self):
-        return self.is_Not, self.code
-
-    @property
-    def lit(self):
-        for k,v in literals_store.items():
-            if v == self.code:
-                return k
-
-    def __invert__(self):
-        is_not = not self.is_Not
-        return Literal(self.code, is_not=is_not)
-
-    def __str__(self):
-        # checking fix later
-        return '~'+str(self.lit) if self.is_Not else str(self.lit)
-        # return type(self).__name__+'( '+ str(self.lit)+ ', ' + str(self.is_Not)+')'
-
-    __repr__ = __str__
-
-    def __eq__(self, other):
-        return (self.is_Not == other.is_Not) and \
-               (self.code == other.code)
-
-    def __hash__(self):
-        h = hash((type(self).__name__,) + self.arg)
-        return h
-
-    def rcall(self, expr):
-        return Literal(self.lit(expr), self.is_Not)
-
-    def __lt__(self, other):
-        sign = -1 if self.is_Not else 1
-        return sign*self.code < sign*other.code
-
-
-def literal_sort_key(lit):
-    _type = type(lit).__name__
-    if _type == 'Literal':
-        return (lit.code, lit.is_Not)
-
-    if _type == 'AND':
-        return (-2, lit.args)
-
-    if _type == 'OR':
-        return (-1, lit.args)
-
-
-class OR(object):
-    def __init__(self, *args):
-        self._args = args
-
-    def __invert__(self):
-        return AND(*[~arg for arg in self._args])
-
-    @property
-    def args(self):
-        return sorted(self._args, key=literal_sort_key)
-
-    def __hash__(self):
-        return hash((type(self).__name__, ) + tuple(self.args))
-
-    def __eq__(self, other):
-        return self.args == other.args
-
-    def __str__(self):
-        s = '(' + ' | '.join([str(arg) for arg in self.args]) + ')'
-        return s
-
-    __repr__ = __str__
-
-
-class AND(object):
-    def __init__(self, *args):
-        self._args = args
-
-    def __invert__(self):
-        return OR(*[~arg for arg in self._args])
-
-    @property
-    def args(self):
-        return sorted(self._args, key=literal_sort_key)
-
-    def __hash__(self):
-        return hash((type(self).__name__, ) + tuple(self.args))
-
-    def __eq__(self, other):
-        return self.args == other.args
-
-    def __str__(self):
-        s = '('+' & '.join([str(arg) for arg in self.args])+')'
-        return s
-
-    __repr__ = __str__
-
-
-def to_NNF(expr):
-    if not isinstance(expr, BooleanFunction):
-        return Literal(expr)
-
-    klass = type(expr).__name__
-    if klass == 'Not':
-        arg = expr.args[0]
-        # attempt 1 just inverting after converting to NNF
-        tmp = to_NNF(arg)
-        return ~tmp
-
-    if klass == 'Or':
-        return OR(*[to_NNF(x) for x in Or.make_args(expr)])
-
-    if klass == 'And':
-        return AND(*[to_NNF(x) for x in And.make_args(expr)])
-
-    if klass == 'Nand':
-        tmp = AND(*[to_NNF(x) for x in expr.args])
-        return ~tmp
-
-    if klass == 'Nor':
-        tmp = OR(*[to_NNF(x) for x in expr.args])
-        return ~tmp
-
-    if klass == 'Xor':
-        cnfs = []
-        for i in range(0, len(expr.args) + 1, 2):
-            for neg in combinations(expr.args, i):
-                clause = [~to_NNF(s) if s in neg else to_NNF(s)
-                          for s in expr.args]
-                cnfs.append(OR(*clause))
-        return AND(*cnfs)
-
-    if klass == 'Xnor':
-        cnfs = []
-        for i in range(0, len(expr.args) + 1, 2):
-            for neg in combinations(expr.args, i):
-                clause = [~to_NNF(s) if s in neg else to_NNF(s)
-                          for s in expr.args]
-                cnfs.append(OR(*clause))
-        return ~AND(*cnfs)
-
-    if klass == 'Implies':
-        L, R = map(to_NNF, expr.args)
-        return OR(~L, R)
-
-    if klass == 'Equivalent':
-        cnfs = []
-        for a, b in zip(expr.args, expr.args[1:]):
-            a = to_NNF(a)
-            b = to_NNF(b)
-            cnfs.append(OR(~a, b))
-        a = to_NNF(expr.args[-1])
-        b = to_NNF(expr.args[0])
-        cnfs.append(OR(~a, b))
-        return AND(*cnfs)
-
-    if klass == 'ITE':
-        L, M, R = map(CNF.to_CNF, expr.args)
-        return AND(OR(~L, M), OR(L, R))
-
-
-def Distribute(expr):
-    if not isinstance(expr, (AND, OR)):
-        tmp = set()
-        tmp.add(frozenset((expr,)))
-        return CNF(tmp)
-
-    klass = type(expr).__name__
-    if klass == 'OR':
-        return CNF.all_or(*[Distribute(arg)
-                            for arg in expr._args])
-
-    if klass == 'AND':
-        return CNF.all_and(*[Distribute(arg)
-                             for arg in expr._args])
-
-
-class CNF(object):
-    def __init__(self, clauses=None):
-        if not clauses:
-            clauses = set()
-        self.clauses = clauses
-
-    def add(self, prop):
-        clauses = CNF.to_CNF(prop).clauses
-        self.clauses |= clauses
-
-    def __str__(self):
-        s = '&'.join(
-            ['(' + '|'.join([str(lit) for lit in clause]) +')'
-            for clause in self.clauses]
-        )
-        return s
-
-    def extend(self, props):
-        for p in props:
-            self.add(p)
-        return self
-
-    def copy(self):
-        return CNF(set(self.clauses))
-
-    def add_clauses(self, clauses):
-        self.clauses |= clauses
-
-    @classmethod
-    def from_prop(cls, prop):
-        res = CNF()
-        res.add(prop)
-        return res
-
-    def __iand__(self, other):
-        self.clauses |= other.clauses
-        return self
-
-    def all_predicates(self):
-        predicates = set()
-        for c in self.clauses:
-            predicates |= {arg.lit for arg in c}
-
-        return predicates
-
-    def _or(self, cnf):
-        clauses = set()
-        # print('in or:',len(self.clauses), len(cnf.clauses))
-        for a,b in product(self.clauses, cnf.clauses):
-            tmp = set(a)
-            for t in b:
-                if ~t in b:
-                    break
-                tmp.add(t)
-            else:
-                clauses.add(frozenset(tmp))
-        # print('or out')
-        # print(clauses)
-        # k = input('move ahead:')
-        return CNF(clauses)
-
-    def _and(self, cnf):
-        # print('in and:', len(self.clauses), len(cnf.clauses))
-        clauses = self.clauses.union(cnf.clauses)
-        # print('out and')
-        # print(clauses)
-        # k = input('move ahead:')
-        return CNF(clauses)
-
-    def _not(self):
-        clss = list(self.clauses)
-        ll = set()
-        for x in clss[-1]:
-            ll.add(frozenset((~x,)))
-        ll = CNF(ll)
-
-        for rest in clss[:-1]:
-            p = set()
-            for x in rest:
-                p.add(frozenset((~x,)))
-            ll = ll._or(CNF(p))
-        return ll
-
-    @classmethod
-    def all_or(cls, *cnfs):
-        b = cnfs[0].copy()
-        for rest in cnfs[1:]:
-            b = b._or(rest)
-        return b
-
-    @classmethod
-    def all_and(cls, *cnfs):
-        b = cnfs[0].copy()
-        for rest in cnfs[1:]:
-            b = b._and(rest)
-        return b
-
-    @classmethod
-    def to_CNF(cls, expr):
-        expr = to_NNF(expr)
-        expr = Distribute(expr)
-        return expr
-
-    @classmethod
-    def CNF_to_cnf(cls, cnf):
-        '''
-        For checking. Remove after everything works
-        '''
-        def remove_literal(arg):
-            return Not(arg.lit) if arg.is_Not else arg.lit
-
-        return And(*(Or(*(remove_literal(arg) for arg in clause)) for clause in cnf.clauses))
-
-
-class EncodedCNF(object):
-    def __init__(self, data=None, encoding=None):
-        if not data and not encoding:
-            data = list()
-            encoding = dict()
-        self.data = data
-        self.encoding = encoding
-        self._symbols = list(encoding.keys())
-
-    def from_cnf(self, cnf):
-        self._symbols = list(cnf.all_predicates())
-        n = len(self._symbols)
-        self.encoding = dict(list(zip(self._symbols, list(range(1, n + 1)))))
-        self.data = [self.encode(clause) for clause in cnf.clauses]
-
-    @property
-    def symbols(self):
-        return self._symbols
-
-    @property
-    def variables(self):
-        return range(1, len(self._symbols) + 1)
-
-    def copy(self):
-        new_data = [set(clause) for clause in self.data]
-        return EncodedCNF(new_data, dict(self.encoding))
-
-    def add_prop(self, prop):
-        cnf = CNF.from_prop(prop)
-        self.add_from_cnf(cnf)
-
-    def add_from_cnf(self, cnf):
-        clauses = [self.encode(clause) for clause in cnf.clauses]
-        self.data += clauses
-
-    def encode_arg(self, arg):
-        literal = arg.lit
-        value = self.encoding.get(literal, None)
-        if value is None:
-            n = len(self._symbols)
-            self._symbols.append(literal)
-            value = self.encoding[literal] = n + 1
-        if arg.is_Not:
-            return -value
-        else:
-            return value
-
-    def encode(self, clause):
-        return {self.encode_arg(arg) if arg is not S.false else arg for arg in clause}
